Adding TestingUtils; refactoring JobConfigFileMonitorTest to use asserts with backoff

+++ gobblin-scheduler/src/test/java/gobblin/scheduler/JobConfigFileMonitorTest.java
@@ -17,9 +17,12 @@ import java.io.FileReader;
 import java.io.FileWriter;
 import java.util.Properties;
 import java.util.Set;
+import java.util.concurrent.Callable;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
@@ -30,6 +33,7 @@ import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ServiceManager;
 
 import gobblin.configuration.ConfigurationKeys;
+import gobblin.testing.TestingUtils;
 
 
 /**
@@ -63,9 +67,15 @@ public class JobConfigFileMonitorTest {
   @Test
   public void testAddNewJobConfigFile()
       throws Exception {
-    Thread.sleep(2000);
-
-    Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 3);
+    final Logger log = LoggerFactory.getLogger("testAddNewJobConfigFile");
+    TestingUtils.assertWithBackoff(new Callable<Boolean>() {
+      @Override
+      public Boolean call() {
+        int numScheduledJobs = JobConfigFileMonitorTest.this.jobScheduler.getScheduledJobs().size();
+        log.debug("numScheduledJobs=" + numScheduledJobs);
+        return numScheduledJobs == 3;
+      }
+    }, 5000, "3 scheduled jobs", log);
 
     // Create a new job configuration file by making a copy of an existing
     // one and giving a different job name
@@ -75,7 +85,14 @@ public class JobConfigFileMonitorTest {
     this.newJobConfigFile = new File(JOB_CONFIG_FILE_DIR, "Gobblin-test-new.pull");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+    TestingUtils.assertWithBackoff(new Callable<Boolean>() {
+      @Override
+      public Boolean call() {
+        int numScheduledJobs = JobConfigFileMonitorTest.this.jobScheduler.getScheduledJobs().size();
+        log.debug("numScheduledJobs=" + numScheduledJobs);
+        return numScheduledJobs == 4;
+      }
+    }, 5000, "4 scheduled jobs", log);
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 4);
@@ -89,6 +106,7 @@ public class JobConfigFileMonitorTest {
   @Test(dependsOnMethods = {"testAddNewJobConfigFile"})
   public void testChangeJobConfigFile()
       throws Exception {
+    final Logger log = LoggerFactory.getLogger("testChangeJobConfigFile");
     Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 4);
 
     // Make a change to the new job configuration file
@@ -97,7 +115,14 @@ public class JobConfigFileMonitorTest {
     jobProps.setProperty(ConfigurationKeys.JOB_COMMIT_POLICY_KEY, "partial");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+    TestingUtils.assertWithBackoff(new Callable<Boolean>() {
+      @Override
+      public Boolean call() {
+        int numScheduledJobs = JobConfigFileMonitorTest.this.jobScheduler.getScheduledJobs().size();
+        log.debug("numScheduledJobs=" + numScheduledJobs);
+        return numScheduledJobs == 4;
+      }
+    }, 5000, "4 scheduled jobs", log);
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 4);
@@ -111,6 +136,7 @@ public class JobConfigFileMonitorTest {
   @Test(dependsOnMethods = {"testChangeJobConfigFile"})
   public void testUnscheduleJob()
       throws Exception {
+    final Logger log = LoggerFactory.getLogger("testUnscheduleJob");
     Assert.assertEquals(this.jobScheduler.getScheduledJobs().size(), 4);
 
     // Disable the new job by setting job.disabled=true
@@ -119,7 +145,15 @@ public class JobConfigFileMonitorTest {
     jobProps.setProperty(ConfigurationKeys.JOB_DISABLED_KEY, "true");
     jobProps.store(new FileWriter(this.newJobConfigFile), null);
 
-    Thread.sleep(2000);
+
+    TestingUtils.assertWithBackoff(new Callable<Boolean>() {
+      @Override
+      public Boolean call() {
+        int numScheduledJobs = JobConfigFileMonitorTest.this.jobScheduler.getScheduledJobs().size();
+        log.debug("numScheduledJobs=" + numScheduledJobs);
+        return numScheduledJobs == 3;
+      }
+    }, 5000, "3 scheduled jobs", log);
 
     Set<String> jobNames = Sets.newHashSet(this.jobScheduler.getScheduledJobs());
     Assert.assertEquals(jobNames.size(), 3);
@@ -131,7 +165,7 @@ public class JobConfigFileMonitorTest {
   @AfterClass
   public void tearDown()
       throws TimeoutException {
-    this.newJobConfigFile.delete();
+    if (null != this.newJobConfigFile) this.newJobConfigFile.delete();
     this.serviceManager.stopAsync().awaitStopped(5, TimeUnit.SECONDS);
   }
 }
